Added VESPA input preparation to riesselman.py.
VESPA input file for single point mutations is easy.
However, it is necessary to calculate all components
of a multiple point mutations separately. That's what
this script does.
Moreover, it is still necessary to add protein name
and an underscore to the beginning of every mutation
with an external tool like awk.